refactor: extract shared formatDurationShort to lib/units
Address Copilot review feedback on PR #69 by removing duplicated
formatDuration helpers from GarminPage and GarminDayActivitiesPopover
and consolidating them into a shared formatDurationShort helper in
src/lib/units.ts (alongside the existing formatDuration). Adds
unit-test coverage for the new helper.
